+++ x/staking/client/cli/utils.go
@@ -6,70 +6,86 @@ import (
 	"fmt"
 	"os"
 
+	"cosmossdk.io/math"
+
 	"github.com/cosmos/cosmos-sdk/codec"
+	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/x/staking/types"
 )
 
 // validator struct to define the fields of the validator
 type validator struct {
-	Amount              string `json:"amount"`
-	From                string `json:"from"`
-	PubKey              string `json:"pubkey"`
-	Moniker             string `json:"moniker"`
-	CommissionRate      string `json:"commission-rate"`
-	CommissionMaxRate   string `json:"commission-max-rate"`
-	CommissionMaxChange string `json:"commission-max-change-rate"`
-	MinSelfDelegation   string `json:"min-self-delegation"`
+	Amount            sdk.Coin
+	PubKey            cryptotypes.PubKey
+	Moniker           string
+	CommissionRates   types.CommissionRates
+	MinSelfDelegation math.Int
 }
 
-func parseValidatorJSON(cdc codec.Codec, path string) (validator, error) {
-	var val validator
+func parseAndValidateValidatorJSON(cdc codec.Codec, path string) (validator, error) {
+	type internalVal struct {
+		Amount              string `json:"amount"`
+		PubKey              string `json:"pubkey"`
+		Moniker             string `json:"moniker"`
+		CommissionRate      string `json:"commission-rate"`
+		CommissionMaxRate   string `json:"commission-max-rate"`
+		CommissionMaxChange string `json:"commission-max-change-rate"`
+		MinSelfDelegation   string `json:"min-self-delegation"`
+	}
 
 	contents, err := os.ReadFile(path)
 	if err != nil {
 		return validator{}, err
 	}
 
-	err = json.Unmarshal(contents, &val)
+	var v internalVal
+	err = json.Unmarshal(contents, &v)
 	if err != nil {
 		return validator{}, err
 	}
 
-	if err := val.validate(); err != nil {
+	if v.Amount == "" {
+		return validator{}, fmt.Errorf("must specify the delegation amount")
+	}
+	amount, err := sdk.ParseCoinNormalized(v.Amount)
+	if err != nil {
 		return validator{}, err
 	}
 
-	return val, nil
-}
-
-// validate checks that the required fields are present in the validator json.
-func (v validator) validate() error {
-	if v.Amount == "" {
-		return fmt.Errorf("must specify amount of coins to bond")
-	}
-	if v.From == "" {
-		return fmt.Errorf("must specify name or address of from key")
-	}
 	if v.PubKey == "" {
-		return fmt.Errorf("must specify the JSON encoded pubkey")
+		return validator{}, fmt.Errorf("must specify the JSON encoded pubkey")
 	}
+	var pk cryptotypes.PubKey
+	if err := cdc.UnmarshalInterfaceJSON([]byte(v.PubKey), &pk); err != nil {
+		return validator{}, err
+	}
+
 	if v.Moniker == "" {
-		return fmt.Errorf("must specify the moniker name")
+		return validator{}, fmt.Errorf("must specify the moniker name")
 	}
-	if v.CommissionRate == "" {
-		return fmt.Errorf("must specify initial commission rate percentage")
+
+	commissionRates, err := buildCommissionRates(v.CommissionRate, v.CommissionMaxRate, v.CommissionMaxChange)
+	if err != nil {
+		return validator{}, err
 	}
-	if v.CommissionMaxRate == "" {
-		return fmt.Errorf("must specify maximum commission rate percentage")
-	}
-	if v.CommissionMaxChange == "" {
-		return fmt.Errorf("must specify maximum commission change rate percentage (per day)")
-	}
+
 	if v.MinSelfDelegation == "" {
-		return fmt.Errorf("minimum self delegation must be a positive integer")
+		return validator{}, fmt.Errorf("must specify minimum self delegation")
 	}
-	return nil
+	minSelfDelegation, ok := sdk.NewIntFromString(v.MinSelfDelegation)
+	if !ok {
+		return validator{}, sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "minimum self delegation must be a positive integer")
+	}
+
+	return validator{
+		Amount:            amount,
+		PubKey:            pk,
+		Moniker:           v.Moniker,
+		CommissionRates:   commissionRates,
+		MinSelfDelegation: minSelfDelegation,
+	}, nil
 }
 
 func buildCommissionRates(rateStr, maxRateStr, maxChangeRateStr string) (commission types.CommissionRates, err error) {